# docs/swagger.yaml
definitions:
  handler.emptyResponse:
    type: object
  handler.errorMsg:
    properties:
      message:
        example: error message
        type: string
    type: object
  handler.successNewSubMsg:
    properties:
      id:
        example: 1
        type: integer
    type: object
  handler.totalCostSum:
    properties:
      sum:
        example: 1000
        type: integer
    type: object
  subsApp.Subscription:
    properties:
      end_date:
        example: 07-2021
        type: string
      id:
        type: integer
      price:
        example: 1000
        type: integer
      service_name:
        example: Web Hosting
        type: string
      start_date:
        example: 06-2020
        type: string
      user_id:
        example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  subsApp.SubscriptionJSON:
    properties:
      end_date:
        example: 07-2021
        type: string
      id:
        type: integer
      price:
        example: 1000
        type: integer
      service_name:
        example: Web Hosting
        type: string
      start_date:
        example: 06-2020
        type: string
      user_id:
        example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  subsApp.UpdSubscription:
    properties:
      end_date:
        example: 07-2021
        type: string
      price:
        example: 1000
        type: integer
      service_name:
        example: Web Hosting
        type: string
      start_date:
        example: 06-2020
        type: string
      user_id:
        example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        type: string
    type: object
info:
  contact: {}
  description: API Server for SubsApp Application
  title: Subs App API
  version: "1.0"
paths:
  /api/v1/subscribe:
    post:
      consumes:
      - application/json
      description: Add a subscription information into database
      operationId: create-subscription
      parameters:
      - description: Subscription main info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/subsApp.Subscription'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.successNewSubMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorMsg'
      summary: CreateSubscription
      tags:
      - subscriptions
  /api/v1/subscribe/{id}:
    delete:
      description: Remove subscription information from database by subscription id
      operationId: del-subscription
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.emptyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorMsg'
      summary: DeleteSubscription
      tags:
      - subscriptions
    get:
      description: Get a subscription information from database by its id
      operationId: get-subscription
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subsApp.SubscriptionJSON'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorMsg'
      summary: GetSubscription
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Update subscription information
      operationId: upd-subscription
      parameters:
      - description: Subscription updated info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/subsApp.UpdSubscription'
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.emptyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorMsg'
      summary: UpdSubscription
      tags:
      - subscriptions
  /api/v1/subscribe/all:
    get:
      description: Get information about all subscriptions from the database
      operationId: get-all-subscriptions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/subsApp.SubscriptionJSON'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorMsg'
      summary: GetAllSubscriptions
      tags:
      - subscriptions
  /api/v1/subscribe/sum:
    get:
      description: Get the total cost of all subscriptions for a selected period,
        filtered by user ID and subscription name
      operationId: get-total-cost-sum-subscriptions
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      - description: Service name
        in: query
        name: service_name
        required: true
        type: string
      - description: Date from
        in: query
        name: date_from
        required: true
        type: string
      - description: Date to
        in: query
        name: date_to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.totalCostSum'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorMsg'
      summary: GetSubsCostSum
      tags:
      - subscriptions
swagger: "2.0"